moltbook-cli/requirements.txt
click>=8.0.0
requests>=2.25.0
pyyaml>=5.4.0
colorama>=0.4.4
python-dotenv>=0.19.0
pytest>=6.0.0
pytest-cov>=2.12.0